Move batch keygen benchmarks to corresponding file
