+++ Cargo.toml
@@ -25,10 +25,6 @@ harness = false
 name = "bench_le"
 harness = false
 
-[[bench]]
-name = "batch"
-harness = false
-
 
 [lib]
 name = "sycret"
