+++ benches/bench_eq.rs
@@ -4,6 +4,7 @@ use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use sycret::eq::*;
 use sycret::stream::{FSSKey, PRG};
 use sycret::utils::MMO;
+use sycret::{eval, keygen};
 
 pub fn eq_keygen(c: &mut Criterion) {
     let mut rng = rand::thread_rng();
@@ -12,6 +13,22 @@ pub fn eq_keygen(c: &mut Criterion) {
     c.bench_function("Eq keygen", |b| {
         b.iter(|| EqKey::generate_keypair(black_box(&mut prg)))
     });
+}
+
+/// op_id = 0 implies Eq key generation
+pub fn eq_batch_keygen(c: &mut Criterion) {
+    let mut keys_a: Vec<u8> = Vec::with_capacity(3105);
+    let mut keys_b: Vec<u8> = Vec::with_capacity(3105);
+    let keys_a_pointer: *mut u8 = keys_a.as_mut_ptr();
+    let keys_b_pointer: *mut u8 = keys_b.as_mut_ptr();
+    let n_values: usize = 5;
+    let n_threads: usize = 6;
+    let op_id: usize = 0;
+    unsafe {
+        c.bench_function("Batch keygen", |b| {
+            b.iter(|| keygen(keys_a_pointer, keys_b_pointer, n_values, n_threads, op_id))
+        });
+    }
 }
 
 pub fn eq_eval(c: &mut Criterion) {
@@ -27,5 +44,5 @@ pub fn eq_eval(c: &mut Criterion) {
     });
 }
 
-criterion_group!(bench_eq, eq_keygen, eq_eval);
+criterion_group!(bench_eq, eq_keygen, eq_eval, eq_batch_keygen);
 criterion_main!(bench_eq);
