+++ benches/bench_le.rs
@@ -4,6 +4,7 @@ use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use sycret::le::*;
 use sycret::stream::{FSSKey, PRG};
 use sycret::utils::MMO;
+use sycret::{eval, keygen};
 
 pub fn le_keygen(c: &mut Criterion) {
     let mut rng = rand::thread_rng();
@@ -12,6 +13,22 @@ pub fn le_keygen(c: &mut Criterion) {
     c.bench_function("Le keygen", |b| {
         b.iter(|| LeKey::generate_keypair(black_box(&mut prg)))
     });
+}
+
+/// op_id = 1 implies Le key generation
+pub fn le_batch_keygen(c: &mut Criterion) {
+    let mut keys_a: Vec<u8> = Vec::with_capacity(4600);
+    let mut keys_b: Vec<u8> = Vec::with_capacity(4600);
+    let keys_a_pointer: *mut u8 = keys_a.as_mut_ptr();
+    let keys_b_pointer: *mut u8 = keys_b.as_mut_ptr();
+    let n_values: usize = 5;
+    let n_threads: usize = 6;
+    let op_id: usize = 1;
+    unsafe {
+        c.bench_function("Batch keygen", |b| {
+            b.iter(|| keygen(keys_a_pointer, keys_b_pointer, n_values, n_threads, op_id))
+        });
+    }
 }
 
 pub fn le_eval(c: &mut Criterion) {
@@ -27,5 +44,5 @@ pub fn le_eval(c: &mut Criterion) {
     });
 }
 
-criterion_group!(bench_le, le_keygen, le_eval);
+criterion_group!(bench_le, le_keygen, le_eval, le_batch_keygen);
 criterion_main!(bench_le);
